pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=23.2.0",
    "aiohttp>=3.9.0",
    "algoliasearch>=4.3.0",
    "algoliasearch-django>=4.0.0",
    "amqp==5.2.0",
    "anthropic>=0.18.0",
    "anyio==4.3.0",
    "arabic-reshaper==3.0.0",
    "asgiref==3.7.2",
    "asn1crypto==1.5.1",
    "attrs==23.1.0",
    "autobahn==23.6.2",
    "automat==22.10.0",
    "autopep8==2.0.4",
    "billiard==4.2.0",
    "cachecontrol==0.14.0",
    "cachetools==5.3.3",
    "cattrs==23.2.3",
    "celery>=5.3.0",
    "certifi==2023.11.17",
    "cffi==1.16.0",
    "channels==4.0.0",
    "chardet==5.2.0",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "click-didyoumean==0.3.0",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "colorama==0.4.6",
    "constantly==23.10.4",
    "cryptography==41.0.7",
    "cssselect2==0.7.0",
    "daphne==4.0.0",
    "defusedxml==0.7.1",
    "diff-match-patch==20230430",
    "django==4.2.5",
    "django-admin-interface==0.28.6",
    "django-colorfield==0.11.0",
    "django-cors-headers==4.2.0",
    "django-crontab==0.7.1",
    "django-filter==23.5",
    "django-import-export==3.3.7",
    "django-jet-reboot==1.3.7",
    "django-js-asset==2.2.0",
    "django-mptt==0.15.0",
    "django-rest-framework==0.1.0",
    "django-schema-graph==3.1.0",
    "django-semantic-admin==0.5.0",
    "django-semantic-forms==0.1.5",
    "djangorestframework==3.14.0",
    "drf-spectacular==0.25.1",
    "et-xmlfile==1.1.0",
    "exceptiongroup==1.2.0",
    "fastapi>=0.104.0",
    "firebase-admin==6.5.0",
    "freetype-py==2.3.0",
    "google-generativeai>=0.3.0",
    "h11==0.14.0",
    "html5lib==1.1",
    "httptools==0.6.1",
    "httpx>=0.25.0",
    "hyperlink==21.0.0",
    "idna==3.6",
    "incremental==22.10.0",
    "inflection==0.5.1",
    "joblib==1.4.2",
    "jsonschema==4.20.0",
    "jsonschema-specifications==2023.11.2",
    "kombu==5.3.5",
    "lxml==4.9.4",
    "markuppy==1.14",
    "msgpack==1.0.8",
    "numpy>=1.24.0",
    "odfpy==1.4.1",
    "openai>=1.3.0",
    "openpyxl==3.1.2",
    "oscrypto==1.3.0",
    "pillow==10.0.1",
    "prompt-toolkit==3.0.43",
    "psycopg2==2.9.9",
    "psycopg2-binary>=2.9.9",
    "pyasn1==0.5.1",
    "pyasn1-modules==0.3.0",
    "pycairo==1.26.0",
    "pycodestyle==2.11.1",
    "pycparser==2.21",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyhanko==0.21.0",
    "pyhanko-certvalidator==0.26.3",
    "pyjwt>=2.8.0",
    "pyopenssl==23.3.0",
    "pyparsing==3.1.2",
    "pypdf==3.17.4",
    "pypng==0.20220715.0",
    "python-bidi==0.4.2",
    "python-dateutil==2.8.2",
    "python-dotenv>=1.0.0",
    "python-json-logger>=2.0.0",
    "python-multipart>=0.0.6",
    "pytz==2023.3.post1",
    "pyyaml==6.0.1",
    "qrcode==7.4.2",
    "redis>=5.0.0",
    "referencing==0.32.0",
    "reportlab==4.0.8",
    "requests==2.31.0",
    "rlpycairo==0.3.0",
    "rpds-py==0.15.2",
    "scikit-learn>=1.8.0",
    "service-identity==24.1.0",
    "shared>=0.0.32",
    "six==1.16.0",
    "sniffio==1.3.1",
    "sqlalchemy>=2.0.0",
    "sqlparse==0.4.4",
    "svglib==1.5.1",
    "tablib==3.5.0",
    "tiktoken>=0.5.0",
    "tinycss2==1.2.1",
    "tomli==2.0.1",
    "twisted==23.10.0",
    "txaio==23.1.1",
    "typing-extensions==4.9.0",
    "tzdata==2023.4",
    "tzlocal==5.2",
    "ua-parser==0.18.0",
    "uritemplate==4.1.1",
    "uritools==4.0.2",
    "urllib3==2.1.0",
    "user-agents==2.2.0",
    "uvicorn[standard]>=0.24.0",
    "vine==5.1.0",
    "watchfiles==0.21.0",
    "wcwidth==0.2.13",
    "webencodings==0.5.1",
    "websockets==12.0",
    "xhtml2pdf==0.2.13",
    "xlrd==2.0.1",
    "xlwt==1.3.0",
    "zope-interface==6.1",
]
